conditionally move concat-util

+++ install.sh
@@ -6,8 +6,10 @@
 ## mu-plugins
 yes | cp -rf ~/mine/vip-go-sandbox/mu-plugins/00-sandbox-helper.php /var/www/wp-content/mu-plugins/.
 
-mv /var/www/wp-content/mu-plugins/http-concat/concat-utils.php /var/www/wp-content/mu-plugins/http-concat/concat-utils.php.original
-yes | cp -rf ~/mine/vip-go-sandbox/mu-plugins/concat-utils.php /var/www/wp-content/mu-plugins/http-concat/concat-utils.php
+if [ -f /var/www/wp-content/mu-plugins/http-concat/concat-utils.php.original ]; then
+  mv /var/www/wp-content/mu-plugins/http-concat/concat-utils.php /var/www/wp-content/mu-plugins/http-concat/concat-utils.php.original
+  yes | cp -rf ~/mine/vip-go-sandbox/mu-plugins/concat-utils.php /var/www/wp-content/mu-plugins/http-concat/concat-utils.php
+fi
 
 ##
 git config --global user.email "kailey.lampert@automattic.com"
